Attempting to match order before pushing onto heap
Since all matched trades include at least one matched trade we check for
a match first and then if the order cannot be completely filled we push
it onto the heap. Appears to provide significant speed up, but is most
effective when many orders are matched. From reading it appears that the
overwhelming majority of orders are never matched, but are cancelled
instead.