Enabled removeBuy(), found and fixed awful bug
The bug was that when completing an executed trade for an incoming sell
we would peek at the buys Q, but then pop from the sells Q. This means
the executed buy would still be in its Q and we would have removed the
next best sell.

This raises serious questions about the effectiveness of the testing I
have been using so far. I don't think more hand crafted unit tests are
the answer. They are too hard to build and the edge cases are too
complex to reason about effectively. The best approach is to build test
cases from exising itch files.