Moved limit and limitset into trade package
This allows trade.Order objects to be aware of these two key
datastructures. Allowing for self removing of Orders from these
datastructures without slow lookups.

This decision locks in the limitheap implementation for the priority
queue. Since the matcher will begin to use the trade.Order's self
removal methods a trade.Order must be a member of a trade.Limit for this
to work properly.

This code is getting complex and difficult to reason about. Some deep
thought may be required. Backing out the choice of limitheap could be
difficult.