Added binary search tree trade/orderbst.go
This data structure is intended to allow us to store Order struct in two
ways.
1: by their Guid
2: by their price, as a collection of limit queues

Rather than passing a *Order into the tree we pass in a *node. If a node
is embedded as a value (not a pointer) in an Order struct, then we can
enjoy good data locality when dealing with moderate collections of
stored orders.

The implementation is surprisingly complex and does not balance itself.
In the future I will convert it into a left leaning red-black tree. I
may revisit the decision to include the queuing (for limit sets) logic
here.

Untested.