Removed duplicate data from trade/Order struct
The Guid and Price fields are now found only in their respective Node
structs. The TraderId and TradeId fields are gone, but can be extracted
from the guid value. This change is made in order to reduce the amount
of expensive writing that needs to be done in order to copy an Order's
values into another Order.