Wrapped up buy/sell/orders into trade/MatchTrees
This allows us to hide a lot of nasty details of the trees, like the
node struct, from the matcher. It also provides the matcher with a very
civilised interface where you push/pop/peek *Orders (as one would expect
to).

The MatchTrees will likely become a vehicle for reintroducing a tree
size counter (or several) and can contain the impact of all of the
upcoming tree experiements we are about to undertake.