Added trade/MatchTrees.Size() function
Shows the combined size of both buy and sell trees. It would be awkward
and expensive to track the individual sizes of the buy and sell trees
because when an order is popped by guid it is hard to determine which
of the two other trees it was also in.