Found issue in rb-tree test. Fixed. Tests fail.
The problem was that the black node balance validation function was
never incrementing the black node count. The count was always 0 and so
was always balanced. Fix this defect. Now it is apparent that the black
nodes are unbalanced and the red-black tree is broken.

Next step is to fix the tree.