Red/Black tree implementation of a guidstore
Now tested and benchmarked. As a first blush it looks like it is
achieving more than 5 million writes per second on my atom
chromebook.

The caveat to that is that if we need to allocate a new node
latency will take a (currently unmeasured) hit. However, I expect to
be able to predict the range of expected guid and we can preallocate
the guidstore's internal structure to eliminate this problem.