Matcher now reponding via chan *trade.Response
Previously matchers used the cbuf package. Right now we are moving to
using chans for all communications in order to build an end to end
system quickly. It is likely that Go's chans will be replaced with a
less flexible, and more performant, set of queues. But that may be well
into the future.