Introducing coordinator package
The coordinator package takes the three components for an end-to-end
matching engine and creates the messaging queues required for them to
communicate. The dispatch package has been removed and is now contained
within the coordinator package.

The three components of an end-to-end matching engine are

Listener
Matcher
Responder

The coordinator is simply a single stateless function which takes these
three components and 'wires' them up with channels. Using this approach
it is very easy to provide mock implementations of any of the components
allowing us to test different components and configurations easily.

Also sneaking in a reduction in the amount of testing of the guidstore
package. This was taking ~2 seconds on my chromebook which is too long.