Order -> OrderNode, OrderData -> Order
The Order struct is now the simple flat data structure representing an
order as it comes in off the network. The complex, adn frankly gnarly,
OrderNode is an order when it resides inside two red-black trees inside
the matching engine itself.

Very happy to get this done, was bugging my daily :)