Package guidstore -> guid
Moved guid helper functions MkGuid(...), GetTraderId(...) and
GetTradeId(...) out of the prioq/ package and into guid/

This was done because the netwk.Listener needs to be able to generate
guids from incoming orders in order to store them in the guidstore. So
we converted the guidstore package into a generic guid package to allow
everyone to work with guids easily.