Merged Order and Response into Message
package trade/ -> msg/

This major refactoring was motivated by the observation that the server
was sending and receiving the same messages.

If we take a BUY order as an example. The order needs to be acknowledged
with a MATCHER_ACK message which contains all the information in the BUY
order. Here we can see that the server is sending messages which look
like orders.

If the BUY order produces a FULL match we can see the opposite.
The FULL match becomes a response that is sent back to the client. But
the client now has to acknowledge the FULL match message. So a
CLIENT_ACK comes back into the server corresponding to the response we
sent previously. Here we can see that the server is receiving messages
which look like responses.

So now all messages look the same. They are distinguished by their Kind
value. The increasingly misnamed trade/ package finally has a sensible
name.