Split MsgKind into MsgKind and MsgRoute
This is done in order to separate the notion that a message is an ORDER
or an ACK etc. with whether it is a BUY, SELL or a CANCEL etc.

For instance a message can be a BUY ORDER, this order will need to be
acked so we can now produce a BUY ACK message. Previously if you were a
BUY then you weren't an ACK and vice versa.