Using real matcher in netwk tests
The netwk package's tests now use the real matcher instead of the mock
we were using previously. This makes the netwk package's unit tests into
a full system integration test. This isn't so bad, but it could confuse
people who expect simple unit tests.

We now have a testing approach which builds up from the simple
components, such as prioq and msg, reusing dependent components for the
tests at each stage. I am comfortable with this approach at this time,
but may in the future move integration style tests into a separate
testing area (but then again I might not).

Also snuck in a COMMAND MsgRoute and moved SHUTDOWN into MsgKind as
shutdown message now has Route=COMMAND and Kind=SHUTDOWN. This was done
so that we had a coherent SERVER_ACK for the shutdown message, which
previously had the kind NO_KIND.