Removed *Data structs in the msg package
These were previously used to aid in the construction of Message
structs. It is clearer to just create the Messages directly. This has
significantly reduced the size of the msg.go file.

We also detected, and fixed, a bug around the wrong IP:Port being used
for SELL reponses. This was undetected by any test and we should ensure
that tests in the netwk package use different IP:Port combos for
different traders.