netwk tests now using unique port for each trader
This detected the bug which I had previously detected, through simple
code reading, and then reintroduced during refactoring. Clearly this is
an important test criteria given the ease with which it was
reintroduced.

As a side note we have seen a number of bugs which appear due to the
symmetry of the BUY and SELL structure in the matcher. Because we are
always dealing with two orders, which are the same bar one detail, it is
common to mix up their fields. Theses bugs can be hard to detect, so it
may be worth considering this in the future.