Adding new MsgStatus field to Message
This is used to capture whether a message is a normal message or a
copy of a message that has caused an error. These copies are called
error messages (for obvious reasons).

There are three possible values for this field

NORMAL
    Not an error message
SENDABLE_ERROR
    An error message which will be logged and sent to the client
    indicated by its IP and Port fields
NOT_SENDABLE_ERROR
    An error message which cannot be sent over the network will only be
    logged

The most obvious instance of an unsendable error would be a message
which arrives into the system with zeroed IP and Port fields.