Message.Valid() now tested for error statuses
We now have two new rules for validity of Messages based on their status

1: NORMAL
    Usual validity rules apply
2: SENDABLE_ERROR
    Message is valid provided the IP and Port fields are not zero
3: NOT_SENDABLE_ERROR
    Always valid

All existing tests are reused and the test messages generated have the
two possible error statuses written to them. We had to split some
message categories so that we knew if a message was expected to have any
missing netwk fields.

Message.IsValid() -> Message.Valid() to be more idiomatic.