Listener/Responder io done via interfaces
Messages being read from and written to the network are using
intermediate interfaces for the purposes. This means we can more easily
swap out network particulars when writing tests.

The most attractive application of this is to be able to more easily
simulate lossy networks in unit tests. We can insert readers and writers
which drop messsages randomly and test the ACK behaviour of the system.

The Listener uses the io.ReadCloser interface which was trivially
swappable with the existing net.UDPConn in the Listener struct. Swapping
a new ReadCloser implementation is done through the constructor.

The Responder uses a custom ipWriter interface which has a local
implementation. This supports a single Write(...) method which takes a
byte slice plus ip:port info. The reason for this is that a UDPConn is
created separately for every response that is written back. This seems
very poor and I strongly expect this approach to be replaced in the
future.