netwk/ Expect(...) now sends CLIENT_ACKs
Whenever a message is consumed in an Expect(...) method we send a
CLIENT_ACK message for the message just consumed to prevent it from
being re-sent. We then wait for a timeout to ensure no new re-sends are
coming. This should prevent some raciness we were seeing in the netwk
testsuit where server re-sends were being picked up and causing test
failures.