Responder now writing to a single ip:port
Previously we attached ip:port info to each message and the responder
would use that to determine where to send the message. This meant that
we needed two classes of errors, errors which could be sent over the
network and errors which could not. Handling this feature proved to be
surprisingly complex.

Now messages have no ip:port info attached and the reponder is
constructed by passing in an io.WriteCloser, to which all messages are
written. This simplifies the error handling, and message validation,
significantly.

There are a number of simple approaches we can take to implement network
communication now.

1: All messages are simply broadcast via UDP - all clients read all
    messages and choose the ones that interest them.
2: A datastructure is created which maps incoming TraderIds to ip:port
    info, the responder uses this to route messages back to individual
    clients
3: The whole system is embedded in another system which communicates via
    channels or some other non-network mechanism. This allows clients
    and matching engine to coexist on one box very easily.

I expect that we will be using UDP broadcast in a production like
environment and using a single 127.0.0.1:<port> UDP connection to act
like a broadcastable network. For super-easy deployment onto VM systems
like Linode we will probably develop a TCP/IP layer wrapping matching
engine and client management into one application.