Message now supports WriteTo/From([]byte)
This pair of methods allow us to serialise/deserialise Messages directly
from a []byte, rather than using the encoding/binary package.

This is useful because it allows us to write undersized datagrams that
come off the network into a Message and then send that message back with
the appropriate error status set.

Additionally we have added specific errors for each kind of error that
can occur. There is a single test for listen.go.