Responder now using map to store unacked messages
This is a significant simplification of the code handling unacked
messages.

When an unacked message is being acked we used to perform a linear
search through a slice of unacked messages looking for the one being
acked. Now we only need to generate the message's guid and remove it
from the map.

The order in which unacked messages are re-sent is no longer well
defined. So we had to modify the unit tests. Where before we would
expect each resent message individually we now expect any one of the
possible resendable messages. For this reason the ExpectNoAck(...)
method was replaced with ExpectOneOf_NoAck(...)