Added cancel tests to testsuite
These tests revealed a bug where CANCEL messages weren't being passed
from the listener to the dispatcher. This was because we were trying to
insert the cancel message's guid into the guidstore and this would
indicate that the message had already been added.

The ExpectEmpty(...) method was removed from the MatchTester interface
in testsuite.go. Previously, networked tests were using timeouts to
determine that there really were no more messages coming, i.e. that the
message queue for that trader was empty. Now we exploit the strong
ordering guarantees on message delivery and we push through another
set of messages for that trader and simply expect on the new messages.
This is referred to in the testsuite as flushing and can be seen in the
cancel and separate stock tests. Using timeouts was also very slow and
removing the ExpectEmpty method has stripped 2-3 seconds off a full
build/test.