Matcher (officially) panics on unexpected messages
This was the previous behaviour too, but now I removed the TODOs to
change this.

The decision to adopt panic in these scenarios is that they represent a
programming failure and are not considered part of normal execution. It
is expected that any erroneous messages are handled by the dispatcher.