client/ with Client, Trader and TraderMaker types
The Client is an empty vessel used to hook into the coordinator package
to provide communication between a large number of traders and a
matching engine. The Client type is very similar to the
coordinator/client type used for testing. It contains two channels to
allow multiple Trader objects to use a single Client for communication.
The intClient channel is for sending messages to a matching engine and
the outOfClient channel is for receiving messages back from the matching
engine. When a Client is created a TraderMaker is also created linked to
that Client.

The intoClient channel also doubles for registering new traders. We use
the same channel to prevent race conditions between registration
messages and messages intended for a matching engine.

The Trader provides a simple interface for submitting trades (buy, sell
and cancel) which are then forwarded to the connected Client. Traders
are made by a TraderMaker.

A TraderMaker is created along side a Client and can be used to create
Traders which are connected to that Client