Renamed channel orders -> appMsgs in coordinator/
Orders was no longer appropriate as a trading client may now be
receiving messages which are not orders through it.

Additional changes:

Responder now panics if it can't handle a message. This may not be the
best long term solution but for now it will give loud and early warning
about any bad messages that make it through.

Message struct now divided (using comments) into header and body
section. We attempted to actually divide these two parts into separate
structs and combine them in Message. But this made initialisation
awkward, so the softer approach of just commenting the distinction is
used. May change in the future.

Renamed responder's handleClientAck(...) -> handleInAck(...) because
previously it was a matching engine handling ACKs from the client, now
it is 'some application' handling ACKs coming IN from the network.