Created trading client for testing
A trading client and matching engine can both use the same messaging
infrastructure provided by the listener/responder/dispatcher. So we can
replace the very complex system of sends and acks in netwk/tutil_test.go
with a simple test client. This removes the raciness that we were seeing
previously in tests because the test client automatically discards
duplicate message sends.

Additional small changes include:

Coordinator.Coordinate(...) now takes a name argument. This is used to
clarify logs printed to stdout from both the trading client and server
simultaneously.

All messages defined in matcher/testsuite.go have had their direction
changed from IN to OUT. Previously we were having to create a message as
it would look when it was sent out from the responder. Now we are
creating messages which look just as they would just after being created
by an app.

netwk/responder's RESEND_MILLIS is now 10 milliseconds. This can be very
short now because both the trading client and matching engine will have
repeat messages automatically de-duplicated.

A pair of unit tests which tested the automatic resends of messages and
the ACK mechanism have been removed. When the trading client is wrapped
in a listener/responder/dispatcher we can no longer mannually perform
acking. There is a TODO to address this. It is very likely that this
specific feature will be subsumed by a set of tests which simulate an
unreliable network.