Removed packaged cbuf
This package was being kept around because it was expected it could be
reused when we came to re-evaluate the use of channels for message
passing. This will not happen for a very long time, if ever, so keeping
this code around is less valuable. Additionally the code is not very
complex and will probably prove to be wildly innapropriate when we start
exploring finely tuned queue implementations.