Removed 'flushing' from matcher testsuite
Previously we had used a technique called 'flushing' to make assertions
about things events which did not happen inside the matching engine. An
example would be a test which asserted that two orders did no result in
a match. Using flushing we would push a few unrelated messages into the
matching engine and that would ensure that no other messages had come
back from the matching engine in the mean time. This approach relied on
the fact that the underlying tests were checking for ACK messages and so
if we pushed though an unrelated message and could only see its ACK
on the message queue we knew that no other messages had been generated.

This doesn't work anymore. Because we have the message handling system
found in the coordinator/ package the ACKs are invisible to the test
system and 'flushing' no longer works. Existing tests using 'flushing'
have been changed to use some other approach with explicit Expect(...)
function calls.

Minor changes include:

Added svr binary to gitignore
"udp" was "upd" in tutil_test.go