Added logging to coordinator/ package
The listener and responder now respect the value of the 'log' value
passed into coordinator.Coordinate(...). The listener will log errors
and the responder will log every message which passes though it's
channel. Currently logging is a simple println(...).

Also added parameter names to the interface methods of msgRunner and
AppMsgRunner.