Process(...) now a function member of AppMsgHelper
This allows different message protocol implementations to pass in
different message processing functions. Right now we have hard coded the
existing Process method, which is now called reliableMsgProcessor.

In the future we will provide different implementations of coordinate
which will use different message protocols. Primarily this has been done
to allow for a low overhead in-memory protocol to be used while the
existing reliable network protocol is further developed.