Removed header fields from msg.Message struct
All the header fields which were required by the reliable messaging
protocol have been removed. The reliable messaging protocol sends and
receives RMessage structs which contain the header fields plus a Message
struct. The application only ever sees msg.Message structs and so any
message protocol which requires additional data must strip this data off
in the listener and add it back on in the responder. The reliable
messaging protocol does exactly this.

This change was made to allow for multiple messaging protocols to be
implemented without having to constantly add fields to the msg.Message
struct.

The msgutil package has been removed. The set and ticker implementations
have been moved to the coordinator/ package and are now private.
guidutil.go has been moved into the msg/ package.