stockManager introduced
We now track the number and kind of stocks that a user owns. The stock
manager was introduced to track this information and also to prevent a
user from selling stock that he/she does not own.

The stock manager maintains two distinct sets of information. A set of
'held' stocks which the user owns. A set of 'to-sell' stocks, which the
owner owns and are currently part of an un-matched/cancelled sell order.

The process for selling stock:

When a sell order is submitted
1: Check that the user has the number of stock he/she wishes to sell.
2: Deduct that number of stocks from the 'held' total and add it to the
   'to-sell' total.

When a sell order is matched (PARTIAL or FULL)
1: Deduct the number of stocks in the match from the 'to-sell' total.

When a sell order is cancelled
1: Deduct that number of stocks from the 'to-sell' total and return it
   to the 'held' total.

The process for buying stock:

When a buy order is matched (PARTIAL or FULL)
1: Add the number of stocks bought to the 'held' total.

Currently a user is given a small number of stocks to begin trading
with.